clean up legacy groups in tests
In the past, we needed the legacy group with PHPUnit < 10 to indicate that a
test was expected to trigger deprecations. With our update to PHPUnit 12 this
is no longer necessary. We now use the group only to skip these tests when
running high deps tests with different major versions. The tests from which
the legacy group is dropped do not handle deprecated Symfony features but
deal with features that users can use to trigger deprecations leveraging our
components. They don't have to be skipped in cross-version test runs.